Add network_interface.py
Add network_interface.py which is a socket server for issuing a stream ID.
Add sample_client.py to check the working of the server.
Notice that original logging level of network_interface.py is DEBUG,
so that the called module 'stream_Id_manager' also have same logging level.
I think it inherits the logging level of 'network_interface.py'.